refactor: changes for some methods
Changing the schema file for texture requires to change this file as well. Now using flatbuffers type instead of common C++ type since we removed the object API from the schema file. Also changed a ref return for a ptr one since most classes use a texture ptr and not a ref one.