multiple changes to the World class
Added the grid as an element of the world. Needed for interaction between engine and gui parts. World will be in charge of adding entities to the grid after the sprite factory added a sprite to the world sprite array. It will also be in charge of calling the EntitySprite destructor to delete the sprite AND the entity to then call the removeEntity method from Grid to just set the ptr to this entity to nullptr.